Reuse `get_merkle_proof` from merkle normal
